Heap

Heap

Functions

Functions

Heap.jl

Heap.indexing — Method

indexing(i, j)::Int64

Giving the depth and the position of the element on that level, returning the index of it in the array

Heap:

0     a
1   b   c
2  d f g h 

Array: [a,b,c,d,e,f,g,h]

To find the position of f: index(i=2, j=2) ightarrow 6

Heap.height — Method

height(i::Int64)::Int64

Get the height of current obj on the heap

Heap.swap — Method

swap(arr::Vector{Int64}, i::Int64, j::Int64)

Swap values of two elements in an array

Heap.convert_index — Method

convert_index(i::Int64)::Int64

Convert python index to julia index

Heap.parent — Method

parent(i::Int64, h::Int64)

Return the parent of current indexing object on the heap

Heap.input — Method

input(io::IO)::Vector{Int64}

Convert an io input into a vector

Heap.max_heapify — Method

max_heapify(arr::Vector{Int64}, index::Int64, depth::Int64)::Vector{Int64}

Runs in $O(lg(n))$ time and is used to maintain the (max) Heap property whenever some node/index i has violated the heap rule (but left subtree, right subtree are each Max Heaps)

Heap.max_heap_insert — Method

max_heap_insert(arr::Vector{Int64},k::Int64)::Vector{Int64}

an insert a new item (and maintain the heap property) in $O(lg(n))$ time.

Same for Heap-Increase-Key

Heap.heap_extract_max — Method

heap_extract_max(arr::Vector{Int64})

Return (and delete) the maximum item of a Heap in $O(lg(n))$ time

Heap.heap_max — Method

heap_max(arr::Vector{Int64})

Returns the max element of a Heap - $heta(1)$ time.

Heap.is_valid_max_heap — Method

is_valid_max_heap(arr::Vector{Int64})::Bool

Check whether an array is a valid max heap

Heap.build_max_heap — Method

build_max_heap(arr::Vector{Int64})::Vector{Int64}

Special one called Build-Max-Heap will run in O(n) time to build a Heap from scratch from an unordered input array.